Move urlJobPopulator, urlJobProcessor into crawler
