+++ package.json
@@ -20,24 +20,20 @@
     "url": "https://github.com/mozilla/prowac/issues"
   },
   "scripts": {
-    "lint": "eslint crawler/**.js dashboard/**.js dataStore/**.js scripts/**.js urlJobPopulator/**.js urlJobProcessor/**.js test/**.js",
+    "lint": "eslint crawler/**.js dashboard/**.js dataStore/**.js scripts/**.js test/**.js",
     "build": "npm-run-all build:**",
       "build:scripts": "babel scripts -d dist/scripts",
       "build:crawler": "babel crawler -d dist/crawler",
       "build:dashboard:js": "babel dashboard/js -d dist/dashboard",
       "build:dashboard:static": "shjs dist/scripts/build-dashboard-static.js",
       "build:datastore": "babel dataStore -d dist/dataStore",
-      "build:urlJobPopulator": "babel urlJobPopulator -d dist/urlJobPopulator",
-      "build:urlJobProcessor": "babel urlJobProcessor -d dist/urlJobProcessor",
     "start": "npm-run-all --parallel start:*",
-      "start:crawler": "echo The crawler is not yet implemented",
+      "start:crawler": "node dist/crawler/main.js",
       "start:dashboard": "echo The dashboard is not yet implemented",
-      "watch:crawler": "npm run build:crawler -- --watch",
+    "watch:crawler": "npm run build:crawler -- --watch",
     "watch:dashboard:js": "npm run build:dashboard:js -- --watch",
     "watch:dashboard:static": "npm run build:dashboard:static -- --watch",
     "watch:datastore": "npm run build:datastore -- --watch",
-    "watch:urlJobPopulator": "npm run build:urlJobPopulator -- --watch",
-    "watch:urlJobProcessor": "npm run build:urlJobProcessor -- --watch",
     "prepublish": "npm run lint && npm run build",
     "test": "npm run build && mocha"
   },
